Fix: create expense form can now handle floats

+++ server/api/groups.js
@@ -78,7 +78,7 @@ router.get('/singleGroup/:groupId/expenses', async (req, res, next) => {
 // POST a new group expense
 router.post('/singleGroup/:groupId/expenses', async (req, res, next) => {
   try {
-    const expenseCost = parseInt(req.body.totalCost)
+    const expenseCost = parseFloat(req.body.totalCost)
     const expenseName = req.body.name
     const newExpense = await Expense.create({
       name: expenseName,
